refactor(xlsx): dedup axis-title extraction; chart_border_width_emu i64→u32
Review follow-up. The axis-title text+run-prop extraction was copy-pasted in
three branches (catAx, scatter is_x_axis, valAx else); fold it into one
extract_axis_title_with_props() helper so the scatter X-axis path stays in
lockstep with catAx. Change chart_border_width_emu from i64 to u32 to match the
other EMU width fields (line/cat/val axis widths). Behavior-preserving (50
parser tests pass; rebuilt WASM is byte-identical).

+++ packages/xlsx/parser/src/types.rs
@@ -751,7 +751,7 @@ pub struct ChartData {
     /// `<c:chartSpace><c:spPr><a:ln@w>` — explicit chart border width in EMU.
     /// None = unset (renderer uses a 1px hairline when a color is present).
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub chart_border_width_emu: Option<i64>,
+    pub chart_border_width_emu: Option<u32>,
 }
 
 /// Generic `<c:manualLayout>` block (used for title, plotArea, legend).
